Change the treatment of cohorts within the same major canopy layer to be completely non-overlapping. Major layer LAI is a biomass-weighted average of cohort LAI within the layer. Modified how a cohort's top canopy position is recorded so that it is retained if lower cohorts die and only a single upper layer remains. Change foliage allocation to happen during the first growing season month, defoliation in the second, and refoliation in the 3rd.  Only defoliated cohorts can refoliate in the 3rd month. An optional parameter (NSCReserve) can limit the withdrawal of NSC for the addition of foliage when a cohort is in a low canopy position.  Default value of 0 sets no limit, consistent with previous behavior. Change CanopyLAImax to float type.

+++ src/ISiteCohorts.cs
@@ -16,7 +16,7 @@ namespace Landis.Extension.Succession.BiomassPnET
         float[] FolResp { get; }
         
         
-        byte CanopyLAImax{get;}
+        float CanopyLAImax{get;}
 
         int AverageAge { get; }
